refactor: Reorganize folders and files for dashboard (DSD-151)
